+++ crates/icu/src/macros/memory.rs
@@ -45,8 +45,12 @@ macro_rules! impl_storable_bounded {
                     is_fixed_size: $is_fixed_size,
                 };
 
-            fn to_bytes(&self) -> ::std::borrow::Cow<[u8]> {
-                ::std::borrow::Cow::Owned(::icu::serialize::serialize(self).unwrap())
+            fn to_bytes(self) -> ::std::borrow::Cow<[u8]> {
+                ::std::borrow::Cow::Owned($crate::serialize::serialize(self).unwrap())
+            }
+
+            fn into_bytes(&self) -> Vec<u8> {
+                $crate::serialize::serialize(self).unwrap()
             }
 
             fn from_bytes(bytes: ::std::borrow::Cow<[u8]>) -> Self {
@@ -66,6 +70,10 @@ macro_rules! impl_storable_unbounded {
 
             fn to_bytes(&self) -> ::std::borrow::Cow<[u8]> {
                 ::std::borrow::Cow::Owned($crate::serialize::serialize(self).unwrap())
+            }
+
+            fn into_bytes(self) -> Vec<u8> {
+                $crate::serialize::serialize(&self).unwrap()
             }
 
             fn from_bytes(bytes: ::std::borrow::Cow<[u8]>) -> Self {